# Optimize sqlite reporter