Goes after #318.
Metrics before/after this PR:

```
FMT:
Saving result: .................... 115ms
Saving result: .................... 10ms

LLVM:
Saving result: .................... 1188ms
Saving result: .................... 10ms
```