@AlexDenisov 2018-02-08
Based on hints from this SO thread:
https://stackoverflow.com/questions/1711631/improve-insert-per-second-performance-of-sqlite